BNH-33 refactor: refactoring according to the PR.
